+++ Frontend/Standard_BlogNote/app/src/main/java/com/example/standardblognote/ui/Components/Item.kt
@@ -55,7 +55,7 @@ fun ItemDocument(
 
     val coroutineScope = rememberCoroutineScope()
     suspend fun HandleCreateNewDocument() {
-        val document = DocumentModel("Untitled", "", "", "", id, uid.toInt())
+        val document = DocumentModel("Untitled", "", "", "", id, uid)
         val res = try {
             RetrofitInstance.api.CreateNewDocument(document)
         } catch (e: HttpException) {
